Add the file for web service
This compressed file includes directories of templates and webs and all other files for creating the database. 
Template directory works together with lib directory; it provides user interface to browse auctions in the database. 
Users are able to log in the system with their username, change current time manually and browse auctions of interest via item id, description, category, price range and status. 
For status, if the current time is before the starting time, then the status of the auction would be "Not Started"; if the current time is after closing time or the item has been bought by someone else, then the status of the auction would be "closed"; otherwise, the status would be "open". Users can add bid on those items have open status. 
If the owner of the auction set a buying price of the item, then once an user bids a price reaching that buying price, then the item belongs to that user and the auction automatically closes. Otherwise, the user who gives the highest bidding price would win after the auction reaches closing time.